\name{NYsite}
\docType{data}
\alias{NYsite}
\title{Position of ozone monitoring sites in New York}
\description{
       This data set provides the Longitude and Latitude  
	of the 29 ozone monitoring sites in the state of New York, USA.
}
\usage{NYsite}
\format{Three columns each containing 29 observations.
\itemize{
	\item 1st col = Site index.
	\item 2nd col = Longitude.
	\item 3rd col = Latitude.
}
}
\examples{
##

data(NYsite)
head(NYsite)

##
}
\source{US EPA}
%\references{
% http://www.epa.gov/
%We do not need anything here. We need See Also NYData
%}
\keyword{datasets}
