\name{spTimer-internal}
\title{Service functions and some undocumented functions for the spTimer library}
\alias{.First.lib}
\alias{.onLoad}
\alias{.onAttach}
\alias{as.forecast.object}
\alias{stat.sum}
\alias{Formula.matrix}
\alias{Formula.coords}
\alias{Gamm}
\alias{Norm}
\alias{Unif}
\alias{PMCC}
\alias{fnc.time}
\alias{initials.checking.ar}
\alias{initials.checking.gp}
\alias{initials.checking.gpp}
\alias{print.spT}
\alias{print.spTpred}
\alias{print.spTprd}
\alias{print.spTfore}
\alias{priors.checking.ar}
\alias{priors.checking.gp}
\alias{priors.checking.gpp}
\alias{spAR.Gibbs}
\alias{spAR.MCMC.Pred}
\alias{spAR.forecast}
\alias{spAR.prediction}
\alias{spGP.Gibbs}
\alias{spGP.MCMC.Pred}
\alias{spGP.forecast}
\alias{spGP.prediction}
\alias{spGPP.Gibbs}
\alias{spGPP.MCMC.Pred}
\alias{spGPP.forecast}
\alias{spGPP.prediction}
\alias{spT.fit.pred}
\alias{spT.hit.false}
\alias{spT.forecast}
\alias{spT.prediction}
\alias{spT.Summary.Stat}
\alias{spT.MCMC.plot}
\alias{spT.MCMC.stat}
\alias{as.mcmc.spT}
\alias{MCMC.plot.obj}
\alias{coef.spT}
\alias{model.frame.spT}
\alias{residuals.spT}
\alias{formula.spT}
\alias{terms.spT}
\alias{model.matrix.spT}
\alias{sp.dimname.fnc}
\alias{tp.dimname.fnc}
\alias{tp}
\alias{sp}
\alias{spT.ObsGridLoc}
\alias{spT.ObsGridData}
\alias{spT.gridTodata}
%\alias{NYgrid}
%\alias{imputation.z}
%\alias{spT.keep.morethan.dist}
%\alias{summary.spT}
%\alias{plot.spT}
\alias{sptruncGP.Gibbs}
\alias{sptruncGPP.Gibbs}
\alias{truncated.fnc}
\alias{prob.below.threshold}
\alias{reverse.truncated.fnc}
\description{Internal spTimer functions}
\keyword{internal}

